drf-01/newsapi/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.2
wrapt==1.12.1
